+++ package/Runtime/BurstDelegateCompiler.cs
@@ -7,7 +7,7 @@ namespace Unity.Burst
         public static unsafe T CompileDelegate<T>(T delegateMethod) where T : class
         {
 #if UNITY_EDITOR
-            int delegateMethodID = Unity.Burst.LowLevel.BurstCompilerService.CompileAsyncDelegateMethod(delegateMethod, "-enable-synchronous-compilation");
+            int delegateMethodID = Unity.Burst.LowLevel.BurstCompilerService.CompileAsyncDelegateMethod(delegateMethod, "--enable-synchronous-compilation");
             void* function = Unity.Burst.LowLevel.BurstCompilerService.GetAsyncCompiledAsyncDelegateMethod(delegateMethodID);
             if (function == null)
                 return delegateMethod;